[Move] utils -> component 폴더 이동
